+++ modules/apps/27-interchain-accounts/controller/types/msgs.go
@@ -36,14 +36,14 @@ func NewMsgRegisterInterchainAccountWithOrdering(connectionID, owner, version st
 }
 
 // NewMsgRegisterInterchainAccount creates a new instance of MsgRegisterInterchainAccount.
-// It uses channeltypes.ORDERED as the default ordering. Breakage in v9.0.0 will allow the ordering to be provided
-// directly. Use NewMsgRegisterInterchainAccountWithOrder to provide the ordering in previous versions.
+// It uses channeltypes.UNORDERED as the default ordering. Breakage in v9.0.0 will allow the ordering to be provided
+// directly. Use NewMsgRegisterInterchainAccountWithOrdering to provide the ordering in previous versions.
 func NewMsgRegisterInterchainAccount(connectionID, owner, version string) *MsgRegisterInterchainAccount {
 	return &MsgRegisterInterchainAccount{
 		ConnectionId: connectionID,
 		Owner:        owner,
 		Version:      version,
-		Ordering:     channeltypes.ORDERED,
+		Ordering:     channeltypes.UNORDERED,
 	}
 }